requirements.txt
gradio>=4.0.0
google-generativeai>=0.7.0
pypdf>=4.0.0
scikit-learn>=1.3.0
